Initial refactor in preparing for WASM (#174)
This PR does a small number of things, which makes it O(_n * m_).

Principally, it is to make room for the WASM bindings generator, with a
`--flavor` ABI Flavor to switch between them.

This necessitated quite a large refactor. Most of this is mechanical; a
lot of changes are files moving and `use` statements being updated.

The upshot of this PR is that we now have a `--flavor wasm` flag in the
`xtask run`, which runs the typescript tests using `tsx`/`nodejs`.

The ReactNative bindings generator calls into a newly refactored
`get_typescript` module which accepts a flavor.

Anticipating that this is not the last time we'll add a switch like
flavor, this is contained in a `SwitchesArg` struct.

+++ scripts/run-tests.sh
@@ -2,14 +2,9 @@
 set -e
 root=.
 
-for test in "${root}"/typescript/tests/*.test.ts ; do
-    echo "Running test $test"
-    cargo xtask run "${test}"
-    echo
-done
-
 declare -a selected_fixtures=()
 declare -a excluded_fixtures=()
+flavor="jsi"
 while (( "$#" )); do
   case "$1" in
     '--fixture'|'-f')
@@ -20,11 +15,39 @@ while (( "$#" )); do
        excluded_fixtures+=("$2")
        shift 2
        ;;
+    '--flavor'|'-F')
+       flavor="$2"
+       shift 2
+       ;;
     *)
        echo "Unknown argument: $1"
        exit 1
        ;;
   esac
+done
+
+supports_flavor() {
+  local fixture="$1"
+  local flavor="$2"
+
+  if [[ "$flavor" == "jsi" ]]; then
+    return 0
+  fi
+
+  local flavor_file="${fixture}/tests/bindings/.supported-flavors.txt"
+  if [[ -f "$flavor_file" ]]; then
+    if grep -q "$flavor" "$flavor_file"; then
+      return 0
+    fi
+  fi
+
+  return 1
+}
+
+for test in "${root}"/typescript/tests/*.test.ts ; do
+    echo "Running test $test"
+    cargo xtask run "${test}" --flavor "$flavor"
+    echo
 done
 
 if [ ${#selected_fixtures[@]} -eq 0 ]; then
@@ -37,13 +60,21 @@ for fixture in ${fixtures} ; do
     if [[ " ${excluded_fixtures[@]} " =~ " ${fixture} " ]]; then
         continue
     fi
-    echo "Running fixture ${fixture}"
+
     # This should all go in either an xtask or into our uniffi-bindgen command.
     # This builds the crate into the target dir.
     fixture_dir="${root}/fixtures/${fixture}"
+
+    if ! supports_flavor "${fixture_dir}" "${flavor}"; then
+      echo "Skipping fixture ${fixture}"
+      continue
+    fi
+
     test_file="${fixture_dir}/tests/bindings/test_${fixture//-/_}.ts"
     config_file="${fixture_dir}/uniffi.toml"
     out_dir="${fixture_dir}/generated"
+
+    echo "Running fixture ${fixture}"
     rm -Rf "${out_dir}" 2>/dev/null
 
     cpp_dir="${out_dir}/cpp"
@@ -58,6 +89,7 @@ for fixture in ${fixtures} ; do
         --ts-dir "${ts_dir}" \
         --toml "${config_file}" \
         --crate "${fixture_dir}" \
+        --flavor "$flavor" \
         "${test_file}"
     echo
 done
